requirements.txt
fastapi
uvicorn
pydantic
requests
openai
python-dotenv
groq
